# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Refactoring",
        "title": "online-judge",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Eutop1a",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:65533",
    "basePath": "/api/v1",
    "paths": {
...
        "/check-picture-code": {
            "post": {
                "description": "检查图片验证码",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verification Code API"
                ],
                "summary": "检查图片验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "图片验证码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "用户登录接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "用户注册接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "multipart/form-data"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
        "/send-code": {
            "post": {
                "description": "发送图片验证码接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verification Code API"
                ],
                "summary": "发送图片验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
        "/send-email-code": {
            "post": {
                "description": "发送邮箱验证码接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verification Code API"
                ],
                "summary": "发送邮箱验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
        "/submissions/code": {
            "post": {
                "description": "提交代码接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Submission API"
                ],
                "summary": "提交代码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "题目id",
                        "name": "problem_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "语言",
                        "name": "language",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "代码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
        "/user-id": {
            "post": {
                "description": "获取用户ID接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "获取用户ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户名不存在",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        },
        "/users/{user_id}": {
            "get": {
                "description": "获取用户详细信息接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "获取用户详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新用户详细信息接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "更新用户详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户密码",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "用户邮箱",
                        "name": "email",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "邮箱验证码",
                        "name": "code",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller._Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "description": "omitempty 字段为空就忽略"
                },
                "msg": {}
            }
        }
    }
}